extract codemirror logic from CodeEditor component
